feat(author): prevent deletion of authors with registered books
- Implement 'existsByAutor' check in LivroRepository
- Add 'OperacaoNaoPermitidaException' to handle restricted operations
- Update 'AutorService' to validate if an author has books before deletion
- Remove 'CascadeType.ALL' from 'Autor' entity to prevent accidental book deletion
- Update 'AutorController' to handle the new exception and return appropriate error responses

+++ src/main/java/io/github/lukastff/libraryapi/service/AutorService.java
@@ -1,7 +1,9 @@
 package io.github.lukastff.libraryapi.service;
 
+import io.github.lukastff.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.lukastff.libraryapi.model.Autor;
 import io.github.lukastff.libraryapi.repository.AutorRepository;
+import io.github.lukastff.libraryapi.repository.LivroRepository;
 import io.github.lukastff.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
@@ -14,10 +16,14 @@ public class AutorService {
 
     private final AutorRepository repository;
     private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository, AutorValidator validator) {
+    public AutorService(AutorRepository repository,
+                        AutorValidator validator,
+                        LivroRepository livroRepository) {
         this.repository = repository;
         this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor) {
@@ -39,6 +45,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
+        if(possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir um autor com livros cadastrados!");
+        }
         repository.delete(autor);
     }
 
@@ -56,5 +65,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 }
